Add back the location & translation + add absolute location option
Initially the old location/rotation was deactivated when the object adding code was replaced by a manual addition of the object to the scene. The location had been replaced by an "add at 3D cursor location". With the current update the TK is added to the 3D cursor location but the user can then change the location relative to this point. Also the rotation was added back.

Add absolute_location prop to allow the user to toggle location from relative to 3D cursor location, to abosolute (relative to world origin).